+++ pkg/numbers/number_test.go
@@ -9,7 +9,8 @@ func TestMin(t *testing.T) {
 	assert.Equal(t, Min(1, 5), 1)
 	assert.Equal(t, Min(22, 5), 5)
 }
+
 func TestMax(t *testing.T) {
-	assert.Equal(t, Max(1, 5), 5)
-	assert.Equal(t, Max(22, 5), 22)
+	assert.Equal(t, Max(1, 5), int64(5))
+	assert.Equal(t, Max(22, 5), int64(22))
 }
